736ceac update e2e test

+++ test/jest/release/creditsDownload.spec.js
@@ -27,7 +27,7 @@ describe('download', () => {
     developmentPage = new DevelopmentPage(client)
   }, 120000)
   afterAll(async () => {
-    await client.deleteSession()
+    client && await client.deleteSession()
   })
   beforeEach(async () => {
     isSuccess = false
@@ -38,7 +38,7 @@ describe('download', () => {
       if (!fs.existsSync(outputPath)) {
         fs.mkdirSync(outputPath, { recursive: true })
       }
-      await client.saveScreenshot(outputPath + `/${expect.getState().currentTestName}.png`)
+      client && await client.saveScreenshot(outputPath + `/${expect.getState().currentTestName}.png`)
     }
   })
   it.skip('delete task', async () => {
@@ -128,7 +128,7 @@ describe('download', () => {
         await homePage.downloadTorrentBtn.waitUntil(async () => {
           if (await homePage.getTask(torrentName) === null) return true
           const statusText = await homePage.getTaskStatus(torrentName, { isLog: false })
-          return !statusText.include('Loading')
+          return !statusText.includes('Loading')
         }, {
           timeout: 60000 * 10,
           timeoutMsg: 'task not start'
